+++ test/resources/language/perl/DependencyA.pm
@@ -1,0 +1,9 @@
+package DependencyA;
+
+use parent Exporter;
+
+@EXPORT = qw($dependency_a);
+
+our $dependency_a = 'foo';
+
+1;