+++ test/resources/language/perl/DependencyB.pm
@@ -1,0 +1,9 @@
+package DependencyB;
+
+use parent Exporter;
+
+@EXPORT = qw($dependency_b);
+
+our $dependency_b = 'bar';
+
+1;